requirements.txt
# Model api free
Groq

# Core Machine Learning and Hugging Face Libraries
transformers
datasets
accelerate
torch

# Manim for video generation
manim

# Data handling and utilities
pandas
numpy

# Agent based
langgraph

# Code checking
astunparse
regex

# For handling environment variables (like API keys)
python-dotenv

# For experimentation and notebooks
# jupyter
# notebook
